[BSK-482] small fixes after the code review
Former-commit-id: f1fc96d4b3a5df5583283a43c4a375acb3435cd8
Former-commit-id: 7ac14e2c5c3ccd963c7b37f29e114ab0400d675d

+++ src/simulation/dynamics/msmForceTorque/msmForceTorque.h
@@ -58,8 +58,8 @@ public:
 private:
     std::vector<std::vector<double>> radiiList;                 //!< vector of MSM sphere radii
     std::vector<std::vector<Eigen::Vector3d>> r_SB_BList;       //!< vector of body-fixed MSM sphere locations
-    long unsigned int numSat;                                   //!< number of satellites
-    long unsigned int numSpheres;                               //!< numer of spheres being modeled
+    unsigned int numSat;                                        //!< number of satellites
+    unsigned int numSpheres;                                    //!< numer of spheres being modeled
     std::vector<double> volt;                                   //!< [V] input voltage for each spacecrat object
     std::vector<Eigen::Vector3d> r_BN_NList;                    //!< [m] list of inertial satellite position vectors
     std::vector<Eigen::MRPd> sigma_BNList;                      //!< [m] list of satellite MRP orientations
